+++ vscode/src/main.ts
@@ -12,7 +12,6 @@ import { InlineChatViewManager } from './chat/InlineChatViewProvider'
 import { MessageProviderOptions } from './chat/MessageProvider'
 import { CODY_FEEDBACK_URL } from './chat/protocol'
 import { VSCodeDocumentHistory } from './completions/context/history'
-import * as CompletionsLogger from './completions/logger'
 import { createProviderConfig } from './completions/providers/createProvider'
 import { registerAutocompleteTraceView } from './completions/tracer/traceView'
 import { InlineCompletionItemProvider } from './completions/vscodeInlineCompletionItemProvider'
@@ -96,7 +95,7 @@ const register = async (
     // Controller for inline Chat
     const commentController = new InlineController(context.extensionPath, telemetryService)
     // Controller for Non-Stop Cody
-    const fixup = new FixupController()
+    const fixup = new FixupController(telemetryService)
     disposables.push(fixup)
     if (TestSupport.instance) {
         TestSupport.instance.fixupController.set(fixup)
@@ -105,7 +104,7 @@ const register = async (
     const editor = new VSCodeEditor({
         inline: commentController,
         fixups: fixup,
-        command: platform.createCommandsController?.(context, localStorage),
+        command: platform.createCommandsController?.(context, localStorage, telemetryService),
     })
 
     // Could we use the `initialConfig` instead?
@@ -185,26 +184,6 @@ const register = async (
         await sidebarChatProvider.executeRecipe(recipe, humanInput)
     }
 
-    const webviewErrorMessenger = async (error: string): Promise<void> => {
-        if (error.includes('rate limit')) {
-            const currentTime: number = Date.now()
-            const userPref = localStorage.get('rateLimitError')
-            // 21600000 is 6h in ms. ex 6 * 60 * 60 * 1000
-            if (!userPref || userPref !== 'never' || currentTime - 21600000 >= parseInt(userPref, 10)) {
-                const input = await vscode.window.showErrorMessage(error, 'Do not show again', 'Close')
-                switch (input) {
-                    case 'Do not show again':
-                        await localStorage.set('rateLimitError', 'never')
-                        break
-                    default:
-                        // Save current time as a reminder stamp in 6 hours
-                        await localStorage.set('rateLimitError', currentTime.toString())
-                }
-            }
-        }
-        sidebarChatProvider.handleError(error)
-    }
-
     const executeFixup = async (
         options: {
             document?: vscode.TextDocument
@@ -229,7 +208,7 @@ const register = async (
             return
         }
 
-        telemetryService.log('CodyVSCodeExtension:fixup')
+        telemetryService.log('CodyVSCodeExtension:fixup:created')
         const provider = fixupManager.getProviderForTask(task)
         return provider.startFix()
     }
@@ -257,18 +236,21 @@ const register = async (
 
             const inlineChatProvider = inlineChatManager.getProviderForThread(comment.thread)
             await inlineChatProvider.addChat(comment.text, false)
-            telemetryService.log('CodyVSCodeExtension:inline-assist:chat')
+            telemetryService.log('CodyVSCodeExtension:chat:submitted', { source: 'inline' })
         }),
         vscode.commands.registerCommand('cody.comment.delete', (thread: vscode.CommentThread) => {
             inlineChatManager.removeProviderForThread(thread)
+            telemetryService.log('CodyVSCodeExtension:inline-assist:deleteButton:clicked')
         }),
         vscode.commands.registerCommand('cody.comment.stop', async (comment: Comment) => {
             const inlineChatProvider = inlineChatManager.getProviderForThread(comment.parent)
             await inlineChatProvider.abortChat()
+            telemetryService.log('CodyVSCodeExtension:abortButton:clicked', { source: 'inline-chat' })
         }),
-        vscode.commands.registerCommand('cody.comment.collapse-all', () =>
-            vscode.commands.executeCommand('workbench.action.collapseAllComments')
-        ),
+        vscode.commands.registerCommand('cody.comment.collapse-all', () => {
+            void vscode.commands.executeCommand('workbench.action.collapseAllComments')
+            telemetryService.log('CodyVSCodeExtension:inline-assist:collapseButton:clicked')
+        }),
         vscode.commands.registerCommand('cody.comment.open-in-sidebar', async (thread: vscode.CommentThread) => {
             const inlineChatProvider = inlineChatManager.getProviderForThread(thread)
             // The inline chat is already saved in history, we just need to tell the sidebar chat to restore it
@@ -277,6 +259,7 @@ const register = async (
             await sidebarChatProvider.setWebviewView('chat')
             // Remove the inline chat
             inlineChatManager.removeProviderForThread(thread)
+            telemetryService.log('CodyVSCodeExtension:inline-assist:openInSidebarButton:clicked')
         }),
         vscode.commands.registerCommand(
             'cody.fixup.new',
@@ -308,27 +291,31 @@ const register = async (
         vscode.commands.registerCommand('cody.interactive.clear', async () => {
             await sidebarChatProvider.clearAndRestartSession()
             await sidebarChatProvider.setWebviewView('chat')
+            telemetryService.log('CodyVSCodeExtension:chatTitleButton:clicked', { name: 'reset' })
         }),
         vscode.commands.registerCommand('cody.focus', () => vscode.commands.executeCommand('cody.chat.focus')),
         vscode.commands.registerCommand('cody.settings.extension', () =>
             vscode.commands.executeCommand('workbench.action.openSettings', { query: '@ext:sourcegraph.cody-ai' })
         ),
-        vscode.commands.registerCommand('cody.history', async () => sidebarChatProvider.setWebviewView('history')),
+        vscode.commands.registerCommand('cody.history', async () => {
+            await sidebarChatProvider.setWebviewView('history')
+            telemetryService.log('CodyVSCodeExtension:chatTitleButton:clicked', { name: 'history' })
+        }),
         vscode.commands.registerCommand('cody.history.clear', async () => {
             await sidebarChatProvider.clearHistory()
         }),
         // Recipes
-        vscode.commands.registerCommand('cody.action.chat', input =>
-            executeRecipeInSidebar('chat-question', true, input)
-        ),
+        vscode.commands.registerCommand('cody.action.chat', async input => {
+            await executeRecipeInSidebar('chat-question', true, input)
+            telemetryService.log('CodyVSCodeExtension:chat:submitted', { source: 'menu' })
+        }),
         vscode.commands.registerCommand(
             'cody.action.fixup',
             (instruction: string, range: vscode.Range): Promise<void> => executeFixup({ instruction, range })
         ),
-        vscode.commands.registerCommand(
-            'cody.action.commands.menu',
-            showDesc => editor.controllers.command?.menu('default', showDesc)
-        ),
+        vscode.commands.registerCommand('cody.action.commands.menu', async () => {
+            await editor.controllers.command?.menu('default')
+        }),
         vscode.commands.registerCommand(
             'cody.action.commands.custom.menu',
             () => editor.controllers.command?.menu('custom')
@@ -340,18 +327,18 @@ const register = async (
             }
             await sidebarChatProvider.executeCustomCommand(title)
         }),
-        vscode.commands.registerCommand('cody.command.explain-code', () =>
-            executeRecipeInSidebar('custom-prompt', true, '/explain')
-        ),
-        vscode.commands.registerCommand('cody.recipe.explain-code', () =>
-            executeRecipeInSidebar('explain-code-detailed')
-        ),
-        vscode.commands.registerCommand('cody.command.generate-unit-test', () =>
-            executeRecipeInSidebar('custom-prompt', true, '/tests')
-        ),
-        vscode.commands.registerCommand('cody.command.generate-docstring', () =>
-            executeRecipeInSidebar('custom-prompt', true, '/docstring')
-        ),
+        vscode.commands.registerCommand('cody.command.explain-code', async () => {
+            await executeRecipeInSidebar('custom-prompt', true, '/explain')
+        }),
+        vscode.commands.registerCommand('cody.command.generate-tests', async () => {
+            await executeRecipeInSidebar('custom-prompt', true, '/test')
+        }),
+        vscode.commands.registerCommand('cody.command.document-code', async () => {
+            await executeRecipeInSidebar('custom-prompt', true, '/doc')
+        }),
+        vscode.commands.registerCommand('cody.command.smell-code', async () => {
+            await executeRecipeInSidebar('custom-prompt', true, '/smell')
+        }),
         vscode.commands.registerCommand('cody.command.inline-touch', () =>
             executeRecipeInSidebar('inline-touch', false)
         ),
@@ -370,6 +357,7 @@ const register = async (
             vscode.env.openExternal(vscode.Uri.parse(CODY_FEEDBACK_URL.href))
         ),
         vscode.commands.registerCommand('cody.welcome', async () => {
+            telemetryService.log('CodyVSCodeExtension:walkthrough:clicked', { page: 'welcome' })
             // Hack: We have to run this twice to force VS Code to register the walkthrough
             // Open issue: https://github.com/microsoft/vscode/issues/186165
             await vscode.commands.executeCommand('workbench.action.openWalkthrough')
@@ -387,10 +375,12 @@ const register = async (
         ),
         vscode.commands.registerCommand('cody.walkthrough.showChat', () => sidebarChatProvider.setWebviewView('chat')),
         vscode.commands.registerCommand('cody.walkthrough.showFixup', () => sidebarChatProvider.setWebviewView('chat')),
-        vscode.commands.registerCommand('cody.walkthrough.showExplain', () =>
-            sidebarChatProvider.setWebviewView('chat')
-        ),
+        vscode.commands.registerCommand('cody.walkthrough.showExplain', async () => {
+            telemetryService.log('CodyVSCodeExtension:walkthrough:clicked', { page: 'showExplain' })
+            await sidebarChatProvider.setWebviewView('chat')
+        }),
         vscode.commands.registerCommand('cody.walkthrough.enableInlineChat', async () => {
+            telemetryService.log('CodyVSCodeExtension:walkthrough:clicked', { page: 'enableInlineChat' })
             await workspaceConfig.update('cody.inlineChat', true, vscode.ConfigurationTarget.Global)
             // Open VSCode setting view. Provides visual confirmation that the setting is enabled.
             return vscode.commands.executeCommand('workbench.action.openSettings', {
@@ -401,48 +391,36 @@ const register = async (
     )
 
     let completionsProvider: vscode.Disposable | null = null
-    if (initialConfig.autocomplete) {
-        completionsProvider = createCompletionsProvider(
-            config,
-            webviewErrorMessenger,
-            completionsClient,
-            statusBar,
-            contextProvider
-        )
-    }
+    disposables.push({ dispose: () => completionsProvider?.dispose() })
+    const setupAutocomplete = (): void => {
+        const config = getConfiguration(vscode.workspace.getConfiguration())
 
-    // Create a disposable to clean up completions when the extension reloads.
-    const disposeCompletions: vscode.Disposable = {
-        dispose: () => {
+        if (!config.autocomplete) {
             completionsProvider?.dispose()
-        },
-    }
-    disposables.push(disposeCompletions)
+            completionsProvider = null
+            if (config.isRunningInsideAgent) {
+                throw new Error(
+                    'The setting `config.autocomplete` evaluated to `false`. It must be true when running inside the agent. ' +
+                        'To fix this problem, make sure that the setting cody.autocomplete.enabled has the value true.'
+                )
+            }
+            return
+        }
 
+        if (completionsProvider !== null) {
+            // If completions are already initialized and still enabled, we
+            // need to reset the completion provider.
+            completionsProvider.dispose()
+        }
+
+        completionsProvider = createCompletionsProvider(config, completionsClient, statusBar, contextProvider)
+    }
     vscode.workspace.onDidChangeConfiguration(event => {
         if (event.affectsConfiguration('cody.autocomplete')) {
-            const config = getConfiguration(vscode.workspace.getConfiguration())
-
-            if (!config.autocomplete) {
-                completionsProvider?.dispose()
-                completionsProvider = null
-                return
-            }
-
-            if (completionsProvider !== null) {
-                // If completions are already initialized and still enabled, we
-                // need to reset the completion provider.
-                completionsProvider.dispose()
-            }
-            completionsProvider = createCompletionsProvider(
-                config,
-                webviewErrorMessenger,
-                completionsClient,
-                statusBar,
-                contextProvider
-            )
+            setupAutocomplete()
         }
     })
+    setupAutocomplete()
 
     // Initiate inline chat when feature flag is on
     if (!initialConfig.inlineChat) {
@@ -478,31 +456,40 @@ const register = async (
 
 function createCompletionsProvider(
     config: Configuration,
-    webviewErrorMessenger: (error: string) => Promise<void>,
     completionsClient: SourcegraphCompletionsClient,
     statusBar: CodyStatusBar,
     contextProvider: ContextProvider
 ): vscode.Disposable {
     const disposables: vscode.Disposable[] = []
 
-    const history = new VSCodeDocumentHistory()
-    const providerConfig = createProviderConfig(config, webviewErrorMessenger, completionsClient)
-    const completionsProvider = new InlineCompletionItemProvider({
-        providerConfig,
-        history,
-        statusBar,
-        getCodebaseContext: () => contextProvider.context,
-        isEmbeddingsContextEnabled: config.autocompleteAdvancedEmbeddings,
-        completeSuggestWidgetSelection: config.autocompleteExperimentalCompleteSuggestWidgetSelection,
-    })
+    const providerConfig = createProviderConfig(config, completionsClient)
+    if (providerConfig) {
+        const history = new VSCodeDocumentHistory()
+        const completionsProvider = new InlineCompletionItemProvider({
+            providerConfig,
+            history,
+            statusBar,
+            getCodebaseContext: () => contextProvider.context,
+            isEmbeddingsContextEnabled: config.autocompleteAdvancedEmbeddings,
+            completeSuggestWidgetSelection: config.autocompleteExperimentalCompleteSuggestWidgetSelection,
+        })
 
-    disposables.push(
-        vscode.commands.registerCommand('cody.autocomplete.inline.accepted', ({ codyLogId, codyLines }) => {
-            CompletionsLogger.accept(codyLogId, codyLines)
-        }),
-        vscode.languages.registerInlineCompletionItemProvider('*', completionsProvider),
-        registerAutocompleteTraceView(completionsProvider)
-    )
+        disposables.push(
+            vscode.commands.registerCommand('cody.autocomplete.inline.accepted', ({ codyLogId, codyCompletion }) => {
+                completionsProvider.handleDidAcceptCompletionItem(codyLogId, codyCompletion)
+            }),
+            vscode.languages.registerInlineCompletionItemProvider('*', completionsProvider),
+            registerAutocompleteTraceView(completionsProvider)
+        )
+    } else if (config.isRunningInsideAgent) {
+        throw new Error(
+            "Can't register completion provider because `providerConfig` evaluated to `null`. " +
+                'To fix this problem, debug why createProviderConfig returned null instead of ProviderConfig. ' +
+                'To further debug this problem, here is the configuration:\n' +
+                JSON.stringify(config, null, 2)
+        )
+    }
+
     return {
         dispose: () => {
             for (const disposable of disposables) {