+++ vscode/webviews/App.tsx
@@ -90,7 +90,7 @@ export const App: React.FunctionComponent<{ vscodeAPI: VSCodeWrapper }> = ({ vsc
                         setEnabledPlugins(message.plugins)
                         break
                     case 'custom-prompts':
-                        setMyPrompts(message.prompts || null)
+                        setMyPrompts(message.prompts?.filter(command => command[1]?.slashCommand) || null)
                         break
                     case 'transcript-errors':
                         setIsTranscriptError(message.isTranscriptError)
@@ -149,6 +149,7 @@ export const App: React.FunctionComponent<{ vscodeAPI: VSCodeWrapper }> = ({ vsc
                     telemetryService={telemetryService}
                     appOS={config?.os}
                     appArch={config?.arch}
+                    uiKindIsWeb={config?.uiKindIsWeb}
                     callbackScheme={config?.uriScheme}
                     onLoginRedirect={onLoginRedirect}
                 />